requirements.txt
duckdb>=0.10.0
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.0.0
httpx>=0.27.0
pandas>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
